requirements.txt
certifi==2019.11.28
Django==2.2.9
django-chunked-upload==2.0.0
django-countries==5.5
django-debug-toolbar==2.1
django-filter==2.2.0
django-sortedm2m==3.0.0
lxml==4.4.2
mysqlclient==1.4.6
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
